Prediction of Drug Molecular Activity(2)/requirements.txt
rdkit==2023.9.1
gensim==4.3.1
pgl>=2.1.5
paddlepaddle>=2.0.0
pahelix
deepchem
chembl_webresource_client
pickle
torch
logging
sklearn
mol2vec==0.1


#paddlepaddle==2.5.0 need protobuf required: >=3.1.0,<=3.20.2 ,but usually installed: 5.x.x